requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
alembic==1.4.3
alt-profanity-check==0.24.0
bidict==0.21.2; python_version >= '3.6'
boto3==1.17.76
botocore==1.20.76
certifi==2020.12.5
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
click==7.1.2
dnspython==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
email-validator==1.1.2
eventlet==0.30.2
faker==8.2.0
flask-cors==3.0.8
flask-login==0.5.0
flask-migrate==2.5.3
flask-socketio==5.0.3
flask-sqlalchemy==2.4.4
flask-wtf==0.14.3
flask==1.1.2
greenlet==1.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
gunicorn==20.0.4
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
itsdangerous==1.1.0
jinja2==2.11.2
jmespath==0.10.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
joblib==1.0.1; python_version >= '3.6'
mako==1.1.3
markupsafe==1.1.1
numpy==1.20.3; python_version >= '3.7'
pyjwt==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
python-engineio==4.2.0
python-socketio==5.3.0
pytz==2021.1
requests==2.25.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
s3transfer==0.4.2
scikit-learn==0.24.0; python_version >= '3.6'
scipy==1.6.3; python_version < '3.10' and python_version >= '3.7'
six==1.15.0
sqlalchemy==1.3.19
text-unidecode==1.3
threadpoolctl==2.1.0; python_version >= '3.5'
twilio==6.59.0
urllib3==1.26.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
werkzeug==1.0.1
wtforms==2.3.3
